src/ai_models/advanced_research_tools.py
```python
import os
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score

AI_MODEL_PATH = os.getenv('AI_MODEL_PATH')
DATA_PATH = os.getenv('DATA_PATH')

class AdvancedResearchTools:
    def __init__(self):
        self.model = RandomForestClassifier()

    def load_data(self):
        data = np.load(os.path.join(DATA_PATH, 'advanced_research_data.npy'))
        labels = np.load(os.path.join(DATA_PATH, 'advanced_research_labels.npy'))
        return data, labels

    def train_model(self, data, labels):
        X_train, X_test, y_train, y_test = train_test_split(data, labels, test_size=0.2, random_state=42)
        self.model.fit(X_train, y_train)
        predictions = self.model.predict(X_test)
        print(f"Model Accuracy: {accuracy_score(y_test, predictions)}")

    def save_model(self):
        np.save(os.path.join(AI_MODEL_PATH, 'advanced_research_model.npy'), self.model)

    def load_model(self):
        self.model = np.load(os.path.join(AI_MODEL_PATH, 'advanced_research_model.npy'))

    def predict(self, data):
        return self.model.predict(data)

if __name__ == "__main__":
    art = AdvancedResearchTools()
    data, labels = art.load_data()
    art.train_model(data, labels)
    art.save_model()
```