src/tests/test_advanced_research_tools.py
import unittest
from src.ai_models.advanced_research_tools import AdvancedResearchTools

class TestAdvancedResearchTools(unittest.TestCase):

    def setUp(self):
        self.advanced_tools = AdvancedResearchTools()

    def test_detect_ghost(self):
        result = self.advanced_tools.detectGhost()
        self.assertIsNotNone(result, "Failed to detect ghost")

    def test_enhance_evp(self):
        result = self.advanced_tools.enhanceEVP()
        self.assertIsNotNone(result, "Failed to enhance EVP")

    def test_detect_paranormal_activity(self):
        result = self.advanced_tools.detectParanormalActivity()
        self.assertIsNotNone(result, "Failed to detect paranormal activity")

    def test_collect_data(self):
        result = self.advanced_tools.collectData()
        self.assertIsNotNone(result, "Failed to collect data")

    def test_process_data(self):
        result = self.advanced_tools.processData()
        self.assertIsNotNone(result, "Failed to process data")

    def test_visualize_data(self):
        result = self.advanced_tools.visualizeData()
        self.assertIsNotNone(result, "Failed to visualize data")

    def test_configure_ai_models(self):
        result = self.advanced_tools.configureAIModels()
        self.assertIsNotNone(result, "Failed to configure AI models")

    def test_configure_interfaces(self):
        result = self.advanced_tools.configureInterfaces()
        self.assertIsNotNone(result, "Failed to configure interfaces")

if __name__ == '__main__':
    unittest.main()